+++ appParsers/ParseDXF.py
@@ -7,11 +7,7 @@
 
 from shapely.geometry import LineString, Point
 from shapely.affinity import rotate
-# Vector was an older name for Vec3; try both for backward compatibility:
-try:
-    from ezdxf.math import Vec3 as ezdxf_vector
-except ImportError:
-    from ezdxf.math import Vector as ezdxf_vector
+from ezdxf.math import Vec3 as ezdxf_vector
 
 from appParsers.ParseFont import *
 from appParsers.ParseDXF_Spline import *